More CLI Commands & Improvements, ReadMe Fixes, Tooltip, Etc.
The CLI has had new commands in 'add' and 'del' for normal commands and 'push' as a debug command, and has new keyword types in 'DEBUG_ERROR', 'SYS' and 'ALERT'.
The CLI has had some bug fixes and error-checking improvements alongside removing italicising as it made text harder to read, and removed alot of underlining to keep underlined text for ID's/Names/Values.
An object-freeing error has been corrected which now allows for the project to be exported successfully.
Tooltip has been added for when CLI url's are hovered over, and can be used for other cases in the future.
The primary font is not imported at 3.0 oversampled to help with the Player Screen text looking blurry when fullscreened.
ReadMe was updated to fix mistakes with outdated info.
Fixed issues with alot of the permanent buttons not being connected making much of the app inoperable.
URL's in the CLI now will open the links  when pressed.
General Manager's 'get_date' now has optional arguments for what to include in the returned string.
Other smaller bug fixes and improvements.