Home Screen, Error Handling Improvements, etc.
The Home screen now features a list of possible recommendations based on the weekday, month and what you were listening to and a random list of 10 albums and songs.
Added setting to toggle if the Home screen is generated or not as it can take time to load.
A lot of smaller bug fixes like fixing a crash when opening a artistless / albumless album / song.
A lot of smaller corrections like fixing some grammar / spelling issues inside the Settings screen.