1.0.3: Volume Control, Networking, CLI Improvements, etc.
The 'build_date' in the Export Data is now in unix time to avoid timezone confusion.
Added a 'ExportFilesTrimmer' addon to remove certain files that are not needed in the exported build.
Modified viewport to be 1920/1080 instead of a smaller ratio of it, this has had the side effect of making all the tooltips smaller.
New Volume Control panel in the Player which has a shortcut to open it, it says the volume, has a mute button and buttons to increment / decrement the volume.
There is now network functionality to check if your version of the app is up to date, this is done on startup with an option to disable it, you can also disable all network functionality with another setting, you can also use the CLI to connect, close the connection and get the latest version.
CLI output text can now be highlighted and copied.
New 'Network' tab in the Profile screen.
Removed 'Solid CLI' from the Accessibility and instead moved it to the Settings as 'CLI Style'.
Made the loading bar on the loading screen run on its own thread so it can animate while everything loads.
Added option to have extra control buttons in the Player which gives you First and Last buttons.
Many settings in the CLI now set their respective user setting when they are set if applicable.
Added 'cli_get_style' option which makes the 'get' command return its content in a tree-like fashion to make it easier to read, or can be left to keep it in the old way.
Added CLI shortcuts and the respective 'shrt' command which can be used to add, delete and execute shortcuts, they can also be executed by typing a hashtag '#' followed by the shortcut name.
Added new network keywords to the CLI alongside a new 'SYS_ALERT', also changed some alert colours.
Switched from CLI-callable commands returned Godot error codes to new custom ones held by the GeneralManager.
Updated the info page from the CLI to include more info.
Made the spellcheck function use String.similarity instead of the custom _get_word_distance, removed the latter as it would be unused.
Multiple new user settings, including one to set if having the app open with prevent your pc from entering power-saving mode / sleep.
More settings are saved when quitting, including if the current song is looped.
Added option to disable the player sending a message when playing a new song.
Many other smaller fixes and improvements.