1.0.4: Social Tab, Alerts, New Debug Commands, etc.
New loading bar; instead of a progress bar it is now a texture which spins using a shader, this is for both the normal and soft loading screens.
New 'Social' tab in the profile which has links for the project GitHub, Itch.io, NZG youtube and for the new NZG discord which has a category for NMP.
New 'Alert' system which will create a popup in the bottom-left to tell the user info, currently this is used only for telling the user their version is outdated if it was checked during startup.
Fixed an issue with the zoom on the playing screen camera which showed the edge of the container when the player was in fullscreen.
New 'create_popup_notification', 'create_popup_box' and 'create_alert' debug commands.
Fix issue with using the 'run_command' command when using call / echo-call.
Certain window-related settings have been removed as they caused issues, now only the screen the window was on and if it was fullscreen or not is saved; when toggling between fullscreen and not the non-fullscreen state is Maximised.
Minor UI-improvements to make things look nicer with the tutorial and popup.
The tutorial now mentions the new discord server.
Fixed issue with the 'keep_screen_on' setting not being set on restarting the app.
When trying to exit the app, added a new 'No, Don't Quit' option, and reworded the options.
Other smaller changes & fixes.